ChebNet, recursive test

+++ model.py
@@ -70,9 +70,16 @@ class CrimePred:
         self.val_dataset = None
         self.test_dataset = None
 
-        self.device = None
         self.model = None
         self.readout = None
+
+        self.device = (
+            "cuda"
+            if torch.cuda.is_available()
+            else "mps"
+            if torch.backends.mps.is_available()
+            else "cpu"
+        )
 
     def build_adj_from_polygons(self, polygon, mode='Queen', knn_n=None):
         from scipy.sparse import csr_matrix
@@ -102,21 +109,23 @@ class CrimePred:
         self.test_dataset = CrimeTorchDataset(test_x_ar, test_y_ar, True)
         return
 
-    def build_model(self, in_channels, out_channels,):
-        self.device = (
-            "cuda"
-            if torch.cuda.is_available()
-            else "mps"
-            if torch.backends.mps.is_available()
-            else "cpu"
-        )
-
-        self.model = TGCN(in_channels=in_channels, out_channels=out_channels,).to(self.device)
+    def build_model(self, in_channels, out_channels, model_name='TGCN', **kwargs):
+        self.reset_model()
+        if model_name == 'TGCN':
+            self.model = ResTGCN(in_channels=in_channels, out_channels=out_channels, ).to(self.device)
+        elif model_name == 'GConvGRU':
+            self.model = ResGConvGRU(in_channels=in_channels, out_channels=out_channels, K=kwargs['K']).to(self.device)
+        else:
+            raise NotImplementedError
         self.readout = torch.nn.Sequential(
             torch.nn.Linear(1, 1, bias=True),
-            # torch.nn.ReLU(),
         ).to(self.device)
         return
+
+    def reset_model(self,):
+        del self.model
+        if torch.cuda.is_available():
+            torch.cuda.empty_cache()
 
     def _crime_seq_collate_fn(self, batch):
         x_full = torch.stack([item[0] for item in batch])
@@ -136,7 +145,7 @@ class CrimePred:
         temporal_batchdata = [next(iter(loader)) for loader in temporal_dataloader]
         return temporal_batchdata  # a list of BatchData
 
-    def train_model(self, batch_size=32, lr=0.001, epoch=1000, dir_cache='./'):
+    def train_model(self, batch_size=32, lr=0.001, epoch=1000, dir_cache='./', test_loc=''):
         import os
         from torch.utils.data import DataLoader
         train_dataloader = DataLoader(
@@ -192,15 +201,16 @@ class CrimePred:
         best_val_loss = 1e8
         for file in os.listdir(dir_cache):
             if file.startswith("model_main_"):
-                best_val_loss = float(file.split("model_main_")[1].split("_")[0].replace('.pth', ''))
+                best_val_loss = float(file.split("model_main_")[1].split("_")[0])
                 break
         if val_loss < best_val_loss:
             for filename in os.listdir(dir_cache):
-                if filename.startswith('model_'):
-                    file_path = os.path.join(dir_cache, filename)
-                    os.remove(file_path)
-            torch.save(self.model, f'{dir_cache}/model_main_{val_loss:>8f}.pth')
-            torch.save(self.readout, f'{dir_cache}/model_readout_{val_loss:>8f}.pth')
+                if filename.startswith('model_') and filename.endswith(".pth"):
+                    if float(filename.split("model_")[1].split("_")[2].replace('.pth', '')) == test_loc:
+                        file_path = os.path.join(dir_cache, filename)
+                        os.remove(file_path)
+            torch.save(self.model, f'{dir_cache}/model_main_{val_loss:>8f}_{test_loc}.pth')
+            torch.save(self.readout, f'{dir_cache}/model_readout_{val_loss:>8f}_{test_loc}.pth')
         return val_loss
 
     def pred_crime_test_set(self, dir_cache):
@@ -212,9 +222,9 @@ class CrimePred:
 
         for filename in os.listdir(dir_cache):
             if filename.startswith('model_main'):
-                model = torch.load(f'{dir_cache}/{filename}', weights_only=False)
+                model = torch.load(f'{dir_cache}/{filename}', weights_only=False).to(self.device)
             if filename.startswith('model_readout'):
-                readout = torch.load(f'{dir_cache}/{filename}', weights_only=False)
+                readout = torch.load(f'{dir_cache}/{filename}', weights_only=False).to(self.device)
 
         with torch.no_grad():
             for _, test_batch in enumerate(test_dataloader):
@@ -227,7 +237,11 @@ class CrimePred:
                     )
                 pred = readout(test_h)
         print(f"Test set predicted.")
-        return pred.cpu().numpy(), test_y.cpu().numpy()
+
+        num_graphs = test_batch[0].batch.max().item() + 1
+        pred_out = np.array([pred.cpu().numpy()[test_batch[0].batch == i] for i in range(num_graphs)])
+        test_y_out = np.array([ test_y.cpu().numpy()[test_batch[0].batch == i] for i in range(num_graphs)])
+        return pred_out, test_y_out
 
 
 class EarlyStopper:
@@ -248,8 +262,11 @@ class EarlyStopper:
                 self.stop = True
 
 
-class TGCN(torch.nn.Module):
-    r"""Obtained from Pytorch Geometric Temporal. Refer to their docs for annotations."""
+class ResTGCN(torch.nn.Module):
+    r"""
+    Obtained from Pytorch Geometric Temporal. Refer to their docs for annotations.
+    Graph Convolutional Gated Recurrent Cell.
+    """
 
     def __init__(
         self,
@@ -259,7 +276,7 @@ class TGCN(torch.nn.Module):
         cached: bool = False,
         add_self_loops: bool = True,
     ):
-        super(TGCN, self).__init__()
+        super(ResTGCN, self).__init__()
 
         self.in_channels = in_channels
         self.out_channels = out_channels
@@ -316,21 +333,24 @@ class TGCN(torch.nn.Module):
         return H
 
     def _calculate_update_gate(self, X, edge_index, edge_weight, H):
-        Z = torch.cat([self.conv_z(X, edge_index, edge_weight), H], axis=1)
-        Z = self.linear_z(Z)
+        X_conv_z = self.conv_z(X, edge_index, edge_weight)
+        Z = torch.cat([X_conv_z, H], axis=1)
+        Z = self.linear_z(Z) + X_conv_z
         Z = torch.sigmoid(Z)
         return Z
 
     def _calculate_reset_gate(self, X, edge_index, edge_weight, H):
-        R = torch.cat([self.conv_r(X, edge_index, edge_weight), H], axis=1)
-        R = self.linear_r(R)
+        X_conv_r = self.conv_r(X, edge_index, edge_weight)
+        R = torch.cat([X_conv_r, H], axis=1)
+        R = self.linear_r(R) + X_conv_r
         R = torch.sigmoid(R)
         return R
 
     def _calculate_candidate_state(self, X, edge_index, edge_weight, H, R):
-        H_tilde = torch.cat([self.conv_h(X, edge_index, edge_weight), H * R], axis=1)
+        H_last = self.conv_h(X, edge_index, edge_weight)
+        H_tilde = torch.cat([H_last, H * R], axis=1)
         H_tilde = self.linear_h(H_tilde)
-        H_tilde = torch.tanh(H_tilde)
+        H_tilde = torch.tanh(H_tilde) + H_last
         return H_tilde
 
     def _calculate_hidden_state(self, Z, H, H_tilde):
@@ -352,50 +372,82 @@ class TGCN(torch.nn.Module):
         return H
 
 
-class TChebCN(torch.nn.Module):
-    r"""Adapted from Pytorch Geometric Temporal."""
+class ResGConvGRU(torch.nn.Module):
+    r"""
+    Obtained from Pytorch Geometric Temporal. Refer to their docs for annotations.
+     Chebyshev Graph Convolutional Gated Recurrent Unit Cell.
+    """
 
     def __init__(
         self,
         in_channels: int,
         out_channels: int,
-        improved: bool = False,
-        cached: bool = False,
-        add_self_loops: bool = True,
+        K: int,
+        normalization: str = "sym",
+        bias: bool = True,
     ):
-        super(TChebCN, self).__init__()
+        super(ResGConvGRU, self).__init__()
 
         self.in_channels = in_channels
         self.out_channels = out_channels
-        self.improved = improved
-        self.cached = cached
-        self.add_self_loops = add_self_loops
-
+        self.K = K
+        self.normalization = normalization
+        self.bias = bias
         self._create_parameters_and_layers()
 
     def _create_update_gate_parameters_and_layers(self):
-        self.conv_z = ChebConv(
+
+        self.conv_x_z = ChebConv(
             in_channels=self.in_channels,
             out_channels=self.out_channels,
-            K=1,
+            K=self.K,
+            normalization=self.normalization,
+            bias=self.bias,
         )
-        self.linear_z = torch.nn.Linear(2 * self.out_channels, self.out_channels)
+
+        self.conv_h_z = ChebConv(
+            in_channels=self.out_channels,
+            out_channels=self.out_channels,
+            K=self.K,
+            normalization=self.normalization,
+            bias=self.bias,
+        )
 
     def _create_reset_gate_parameters_and_layers(self):
-        self.conv_r = ChebConv(
+
+        self.conv_x_r = ChebConv(
             in_channels=self.in_channels,
             out_channels=self.out_channels,
-            K=1,
+            K=self.K,
+            normalization=self.normalization,
+            bias=self.bias,
         )
-        self.linear_r = torch.nn.Linear(2 * self.out_channels, self.out_channels)
+
+        self.conv_h_r = ChebConv(
+            in_channels=self.out_channels,
+            out_channels=self.out_channels,
+            K=self.K,
+            normalization=self.normalization,
+            bias=self.bias,
+        )
 
     def _create_candidate_state_parameters_and_layers(self):
-        self.conv_h = ChebConv(
+
+        self.conv_x_h = ChebConv(
             in_channels=self.in_channels,
             out_channels=self.out_channels,
-            K=1,
+            K=self.K,
+            normalization=self.normalization,
+            bias=self.bias,
         )
-        self.linear_h = torch.nn.Linear(2 * self.out_channels, self.out_channels)
+
+        self.conv_h_h = ChebConv(
+            in_channels=self.out_channels,
+            out_channels=self.out_channels,
+            K=self.K,
+            normalization=self.normalization,
+            bias=self.bias,
+        )
 
     def _create_parameters_and_layers(self):
         self._create_update_gate_parameters_and_layers()
@@ -407,22 +459,26 @@ class TChebCN(torch.nn.Module):
             H = torch.zeros(X.shape[0], self.out_channels).to(X.device)
         return H
 
-    def _calculate_update_gate(self, X, edge_index, edge_weight, H):
-        Z = torch.cat([self.conv_z(X, edge_index, edge_weight), H], axis=1)
-        Z = self.linear_z(Z)
+    def _calculate_update_gate(self, X, edge_index, edge_weight, H, lambda_max):
+        Z = self.conv_x_z(X, edge_index, edge_weight, lambda_max=lambda_max)
+        Z = Z + self.conv_h_z(H, edge_index, edge_weight, lambda_max=lambda_max)
         Z = torch.sigmoid(Z)
         return Z
 
-    def _calculate_reset_gate(self, X, edge_index, edge_weight, H):
-        R = torch.cat([self.conv_r(X, edge_index, edge_weight), H], axis=1)
-        R = self.linear_r(R)
+    def _calculate_reset_gate(self, X, edge_index, edge_weight, H, lambda_max):
+        R = self.conv_x_r(X, edge_index, edge_weight, lambda_max=lambda_max)
+        R = R + self.conv_h_r(H, edge_index, edge_weight, lambda_max=lambda_max)
         R = torch.sigmoid(R)
         return R
 
-    def _calculate_candidate_state(self, X, edge_index, edge_weight, H, R):
-        H_tilde = torch.cat([self.conv_h(X, edge_index, edge_weight), H * R], axis=1)
-        H_tilde = self.linear_h(H_tilde)
-        H_tilde = torch.tanh(H_tilde)
+    def _calculate_candidate_state(self, X, edge_index, edge_weight, H, R, lambda_max):
+        # H_tilde = self.conv_x_h(X, edge_index, edge_weight, lambda_max=lambda_max)
+        # H_tilde = H_tilde + self.conv_h_h(H * R, edge_index, edge_weight, lambda_max=lambda_max)
+        # H_tilde = torch.tanh(H_tilde)
+        H_tilde_1 = self.conv_x_h(X, edge_index, edge_weight, lambda_max=lambda_max)
+        H_tilde_2 = self.conv_h_h(H * R, edge_index, edge_weight, lambda_max=lambda_max)
+        H_tilde = H_tilde_1 + H_tilde_2
+        H_tilde = torch.tanh(H_tilde) + H_tilde_1 + H_tilde_2
         return H_tilde
 
     def _calculate_hidden_state(self, Z, H, H_tilde):
@@ -435,30 +491,41 @@ class TChebCN(torch.nn.Module):
             edge_index: torch.LongTensor,
             edge_weight: torch.FloatTensor = None,
             H: torch.FloatTensor = None,
+            lambda_max: torch.Tensor = None,
     ) -> torch.FloatTensor:
         H = self._set_hidden_state(X, H)
-        Z = self._calculate_update_gate(X, edge_index, edge_weight, H)
-        R = self._calculate_reset_gate(X, edge_index, edge_weight, H)
-        H_tilde = self._calculate_candidate_state(X, edge_index, edge_weight, H, R)
+        Z = self._calculate_update_gate(X, edge_index, edge_weight, H, lambda_max)
+        R = self._calculate_reset_gate(X, edge_index, edge_weight, H, lambda_max)
+        H_tilde = self._calculate_candidate_state(X, edge_index, edge_weight, H, R, lambda_max)
         H = self._calculate_hidden_state(Z, H, H_tilde)
         return H
 
 
 class CrimePredTuner(CrimePred):
-    def __init__(self, name_prefix='', storage_path="sqlite:///optuna_study.db",):
+    def __init__(self, model_name, storage_path="sqlite:///optuna_study.db", model_save='./'):
         import datetime
         super().__init__()
         self.storage_path = storage_path
-        self.study_name = f"{name_prefix}{datetime.datetime.now().strftime("%m%d%H%M")}"
+        self.model_name = model_name
+        self.study_name = f"{model_name}_{datetime.datetime.now().strftime("%m%d%H%M")}"
+        self.model_save = model_save
 
     def objective(self, trial):
-        batch_size = trial.suggest_categorical("batch_size", [16, 32, 64])
-        lr = trial.suggest_float("lr", 1e-4, 1e-2, log=True)
-        self.build_model(1, 1, )
-        val_loss = self.train_model(batch_size=batch_size, lr=lr,)
+        if self.model_name == 'TGCN':
+            batch_size = trial.suggest_categorical("batch_size", [8, 16, 32, 64])
+            lr = trial.suggest_float("lr", 1e-4, 1e-2, log=True)
+            self.build_model(1, 1, model_name=self.model_name)
+        elif self.model_name == 'GConvGRU':
+            batch_size = 8  # trial.suggest_categorical("batch_size", [8, 16])
+            lr = trial.suggest_float("lr", 1e-3, 1e-2, log=True)
+            K = trial.suggest_int("K", 1, 3)
+            self.build_model(1, 1, model_name=self.model_name, K=K)
+        else:
+            raise NotImplementedError
+        val_loss = self.train_model(batch_size=batch_size, lr=lr, dir_cache=self.model_save)
         return val_loss
 
-    def run_study(self, n_trials=50):
+    def run_study(self, n_trials=50,):
         import optuna
         study = optuna.create_study(
             direction="minimize", storage=self.storage_path, study_name=self.study_name
@@ -466,4 +533,5 @@ class CrimePredTuner(CrimePred):
         study.optimize(self.objective, n_trials=n_trials)
         print("Best hyperparameters:", study.best_params)
         return study.best_params
+
